rename directory for netlookup
