+++ NetPTN.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetPing", "NetPing\NetPing.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetTrace", "NetTrace\NetTrace.csproj", "{152C2300-5AE1-4551-BFEE-275266AEFC45}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetLookup", "NetNsLookup\NetLookup.csproj", "{E22ED323-7762-4AAA-805A-A3D51057B127}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetLookup", "NetLookup\NetLookup.csproj", "{BFB4249F-0C48-46C3-816C-2F6B4E3360A0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{152C2300-5AE1-4551-BFEE-275266AEFC45}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{152C2300-5AE1-4551-BFEE-275266AEFC45}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{152C2300-5AE1-4551-BFEE-275266AEFC45}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E22ED323-7762-4AAA-805A-A3D51057B127}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E22ED323-7762-4AAA-805A-A3D51057B127}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E22ED323-7762-4AAA-805A-A3D51057B127}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E22ED323-7762-4AAA-805A-A3D51057B127}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BFB4249F-0C48-46C3-816C-2F6B4E3360A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BFB4249F-0C48-46C3-816C-2F6B4E3360A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BFB4249F-0C48-46C3-816C-2F6B4E3360A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BFB4249F-0C48-46C3-816C-2F6B4E3360A0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
